clean and split FileExplorer file
